Revert "[gtest] conversion for code coverage tests (#2580)"
This reverts commit c5a2384dc0f29682ed51aeccf9b981dbdf7e058f.
